requirements.txt
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
colorama==0.4.6
tqdm==4.56.0
whisper
wavio
sounddevice
pydub
